# Conflicts:
#	target/classes/DataBase/DetailsBattle.json
#	target/classes/DataBase/Event.json
#	target/classes/DataBase/Games.json
#	target/classes/DataBase/Message.json
#	target/classes/DataBase/Player.json
#	target/classes/DataBase/Suggestion.json
#	target/classes/DataBase/User.json

+++ src/main/java/Client/View/ConfirmPasswordControllerForDeleteAccount.java
@@ -1,8 +1,6 @@
 package Client.View;
 
-import Server.Controller.CompetencyController.Existence;
-import Server.Controller.Exception.Plato.ExistPlayerException;
-import Server.Controller.PlayerController.PlayerGeneralController;
+import Client.DataLoader;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -21,11 +19,9 @@ import java.io.IOException;
 import java.net.URL;
 
 public class ConfirmPasswordControllerForDeleteAccount {
-    protected static PlayerGeneralController playerGeneralController = new PlayerGeneralController();
+    private static final DataLoader dataLoader = new DataLoader();
 
-
-    private  String username = LoginController.getUsername();
-
+    private String username = LoginController.getUsername();
     protected static String confirm = "false";
 
     public static String getConfirm() {
@@ -53,29 +49,31 @@ public class ConfirmPasswordControllerForDeleteAccount {
 
 
     @FXML
-    private void close(ActionEvent event){
-        Stage stage = (Stage)btnBack.getScene().getWindow();
+    private void close(ActionEvent event) {
+        Stage stage = (Stage) btnBack.getScene().getWindow();
         stage.close();
     }
+
     @FXML
-    private void setBtnConfirm(ActionEvent event) throws IOException, ExistPlayerException {
+    private void setBtnConfirm(ActionEvent event) throws IOException {
         playMouseSound();
-        String password = txtPassword.getText();
-        setConfirm(Existence.checkPasswordForView(getUsername(),password));
-        if (getConfirm().equalsIgnoreCase("false")){
-            showError();
+        if (txtPassword.getText().isEmpty()){
+            System.err.println("Field is Empty");
+            return;
         }
-        else {
-            String[] user = playerGeneralController.showBasicInformation(getUsername()).split("\\$");
-//            File file = new File(user[6]);
-            deleteDirectory(new File("src"+File.separator+"main"+File.separator+"resources"+File.separator+"Users"+File.separator+getUsername()));
-            playerGeneralController.deleteUser(getUsername());
+        setConfirm(dataLoader.confirmPassWord(getUsername(),txtPassword.getText()));
+        if (getConfirm().equalsIgnoreCase("false")) {
+            showError();
+        } else {
+            deleteDirectory(new File("src" + File.separator + "main" + File.separator + "resources" + File.separator + "Users" + File.separator + getUsername()));
+            dataLoader.deletePlayer(getUsername());
             setConfirm("false");
             System.exit(1);
             //todo Check this
         }
 
     }
+
     @FXML
     private void deleteDirectory(File fileForDelete) throws IOException {
         playMouseSound();
@@ -92,7 +90,8 @@ public class ConfirmPasswordControllerForDeleteAccount {
         stage.setScene(scene);
         stage.show();
     }
-    public void playMouseSound(){
+
+    public void playMouseSound() {
         File file = new File("src\\main\\resources\\Sound\\Click.mp3");
         Media media = new Media(file.toURI().toString());
         MediaPlayer mediaPlayer = new MediaPlayer(media);
